Update to 2.2.0.1
Bugfix release

It fixes a minor vulnerability with a mature shapelib contained in our codebase.
The vulnerability addresses only the plugin Importshp, which is used to import shape files (SHP/SHX/DBF).
Shape files are used in surveying and so do not affect the most users.
As this is probably not a widely used plugin, the fix was just to remove the plugin.

**Security Fixes:**
- CVE-2023-30259

Signed-off-by: Algent Albrahimi <algent@protonmail.com>